Change user_id Int -> Long

+++ src/main/kotlin/jp/ac/titech/cs/se/refactorhub/controllers/user/UserController.kt
@@ -14,7 +14,7 @@ class UserController(
 ) {
 
     @GetMapping("/{id}")
-    fun get(@PathVariable("id") id: Int) = ResponseEntity.ok(userService.get(id))
+    fun get(@PathVariable("id") id: Long) = ResponseEntity.ok(userService.get(id))
 
     @GetMapping("/me")
     fun me() = ResponseEntity.ok(userService.me())
